chapter6/ptl/requirements.txt
#torch>=1.6,<1.9
torchvision
torchmetrics>=0.3
pytorch-lightning>=1.3
ipykernel

